Agno_with_mercury2c/requirements.txt
agno
openai
python-dotenv
yfinance
duckduckgo-search